PlayerEvasionSpeedのテストと実装

+++ Assets/Scripts/ValueObjects/Player/PlayerMoveSpeed.cs
@@ -1,9 +1,9 @@
 public class PlayerMoveSpeed : Speed {
 
-    private PlayerMoveSpeed(int value) : base(value, 0, 10) {
+    private PlayerMoveSpeed(float value) : base(value, 0f, 10f) {
     }
 
-    public static PlayerMoveSpeed Of(int value) {
+    public static PlayerMoveSpeed Of(float value) {
         return new PlayerMoveSpeed(value);
     }
