+++ Assets/Scripts/ValueObjects/Speed.cs
@@ -1,8 +1,8 @@
 using KataokaLib.ValueObject;
 
-public abstract class Speed : SingleValueObject<int> {
+public abstract class Speed : SingleValueObject<float> {
 
-    protected Speed(int value, int MIN, int MAX) : base(value, MIN, MAX) {
+    protected Speed(float value, float MIN, float MAX) : base(value, MIN, MAX) {
     }
 
 }
